% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{arena_results}
\alias{arena_results}
\title{Test Dataset with Arena Results}
\format{A data frame with 2280 rows and 4 variables:
\describe{
  \item{Scenario}{The Scenario Name}
  \item{Statistic}{The Statistic's description}
  \item{Replication}{The Replication Number}
  \item{Value}{The numeric value of the statistic within the replication and scenario}
}}
\usage{
arena_results
}
\description{
A dataset containing test data from an Arena simulation model
}
\keyword{datasets}
