% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_confint}
\alias{plot_confint}
\title{Confidence Interval Plot}
\usage{
plot_confint(sim_results, response_variable)
}
\arguments{
\item{sim_results}{The data.frame generated by get_simulation_results()}

\item{response_variable}{A character string indicating the Statistic to be plotted.}
}
\value{
a confidence interval plot using ggplot2.
}
\description{
Plots the confidence interval for a response variable, across different simulated scenarios.
}
\examples{
library(arena2r)

plot_confint(arena_results, "Entity 1.WaitTime")
}
