% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runArenaApp}
\alias{runArenaApp}
\title{Run Arena App}
\usage{
runArenaApp()
}
\description{
This function will launch a Shiny App allowing you to analyse Arena results without writing R code.
}
