% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_statistics_summary}
\alias{get_statistics_summary}
\title{Get Statistics Summary}
\usage{
get_statistics_summary(sim_results, confidence = 0.95)
}
\arguments{
\item{sim_results}{The data.frame generated by get_simulation_results()}

\item{confidence}{The confidence of the CI}
}
\value{
a data.frame with a summary for every Statistic
}
\description{
Makes a summary table to every statistic available
}
\examples{
library(arena2r)

statistics_summary = get_statistics_summary(arena_results)
head(statistics_summary)
}
